+++ plugin/src/__tests__/modifyXcodeProject-test.ts
@@ -2,6 +2,14 @@ import { WarningAggregator } from '@expo/config-plugins';
 
 import { modifyExistingXcodeBuildScript } from '../withSentryIOS';
 
+jest.mock('@expo/config-plugins', () => {
+  const plugins = jest.requireActual('@expo/config-plugins');
+  return {
+    ...plugins,
+    WarningAggregator: { addWarningIOS: jest.fn(), addWarningAndroid: jest.fn() },
+  };
+});
+
 const buildScriptWithoutSentry = {
   shellScript:
     '"export NODE_BINARY=node\\n../node_modules/react-native/scripts/react-native-xcode.sh"',
@@ -64,6 +72,6 @@ describe('Configures iOS native project correctly', () => {
 
   it(`Warns to file a bug report if build script isn't what we expect to find`, () => {
     modifyExistingXcodeBuildScript(buildScriptWeDontExpect);
-    expect(console.warn).toHaveBeenCalled();
+    expect(WarningAggregator.addWarningIOS).toHaveBeenCalled();
   });
 });